requirements.txt
flask==3.1.0
openai==1.65.0
PyGithub==2.5.0
python-dotenv==1.0.1
